Added Game Design Document
This document outlines the objectives for the game, along with what
needs to be developed in terms of music, art and programming structure.
Any edits must be made clear in the main Commit Message above, with
appropriate description below of what was changed, why, and how this
will effect gameplay (if applicable).