Definitely working RunMe.jar
Tested on desktop and all :P Up to date with XMLReader. Need to paste
the XML into browser searchbar (removes formatting) then into the string
called XML in the XMLReader class. It will search for the sign with the
correct ID then print out the text held between <text> and </text>.
XMLReader can be extended easily to read other XMLs or read more data
from them (character talk?) TODO work on a way of formatting the output
string to ensure it fits on the on-screen textbox.